Consolidate file
